chore!: drop support for obsolete "hapi" package
Closes: #2691
