+++ lib/instrumentation/index.js
@@ -61,7 +61,7 @@ var MODULES = [
   'generic-pool',
   'graphql',
   'handlebars',
-  ['hapi', '@hapi/hapi'],
+  '@hapi/hapi',
   'http',
   'https',
   'http2',
