+++ lib/instrumentation/modules/@hapi/hapi.js
@@ -8,39 +8,29 @@
 
 var semver = require('semver');
 
-var shimmer = require('../shimmer');
+var shimmer = require('../../shimmer');
 
 var onPreAuthSym = Symbol('ElasticAPMOnPreAuth');
 
 module.exports = function (hapi, agent, { version, enabled }) {
-  if (!enabled) return hapi;
+  if (!enabled) {
+    return hapi;
+  }
+  if (!semver.satisfies(version, '>=17.9.0 <22.0.0')) {
+    agent.logger.debug('@hapi/hapi@%s not supported, skipping', version);
+    return hapi;
+  }
 
   agent.setFramework({ name: 'hapi', version, overwrite: false });
 
-  if (!semver.satisfies(version, '>=9.0.0')) {
-    agent.logger.debug('hapi version %s not supported - aborting...', version);
-    return hapi;
-  }
-  const isHapiGte17 = semver.satisfies(version, '>=17');
-
-  agent.logger.debug('shimming hapi.Server.prototype.initialize');
-
-  if (isHapiGte17) {
-    shimmer.massWrap(hapi, ['Server', 'server'], function (orig) {
-      return function (options) {
-        var res = orig.apply(this, arguments);
-        patchServer(res);
-        return res;
-      };
-    });
-  } else {
-    shimmer.wrap(hapi.Server.prototype, 'initialize', function (orig) {
-      return function () {
-        patchServer(this);
-        return orig.apply(this, arguments);
-      };
-    });
-  }
+  agent.logger.debug('shimming hapi.Server, hapi.server');
+  shimmer.massWrap(hapi, ['Server', 'server'], function (orig) {
+    return function (options) {
+      var res = orig.apply(this, arguments);
+      patchServer(res);
+      return res;
+    };
+  });
 
   function patchServer(server) {
     // Hooks that are always allowed
@@ -52,40 +42,14 @@ module.exports = function (hapi, agent, { version, enabled }) {
       agent.logger.debug('unable to enable hapi error tracking');
     }
 
-    // Prior to hapi 17, when the server has no connections we can't make
-    // connection lifecycle hooks (in hapi 17+ the server always has
-    // connections, though the `server.connections` property doesn't exists,
-    // so this if-statement wont fire)
-    var conns = server.connections;
-    if (conns && conns.length === 0) {
-      agent.logger.debug(
-        'unable to enable hapi instrumentation on connectionless server',
-      );
-      return;
-    }
-
-    // Hooks that are only allowed when the hapi server has connections
-    // (with hapi 17+ this is always the case)
-    if (typeof server.ext === 'function') {
-      server.ext('onPreAuth', onPreAuth);
-      server.ext('onPreResponse', onPreResponse);
-      if (agent._conf.captureBody !== 'off') {
-        server.ext('onPostAuth', onPostAuth);
-      }
-    } else {
-      agent.logger.debug('unable to enable automatic hapi transaction naming');
+    server.ext('onPreAuth', onPreAuth);
+    server.ext('onPreResponse', onPreResponse);
+    if (agent._conf.captureBody !== 'off') {
+      server.ext('onPostAuth', onPostAuth);
     }
   }
 
   function attachEvents(emitter) {
-    if (!isHapiGte17) {
-      emitter.on('request-error', function (request, error) {
-        agent.captureError(error, {
-          request: request.raw && request.raw.req,
-        });
-      });
-    }
-
     emitter.on('log', function (event, tags) {
       captureError('log', null, event, tags);
     });
@@ -150,7 +114,7 @@ module.exports = function (hapi, agent, { version, enabled }) {
       }
     }
 
-    return isHapiGte17 ? reply.continue : reply.continue();
+    return reply.continue;
   }
 
   function onPostAuth(request, reply) {
@@ -158,7 +122,7 @@ module.exports = function (hapi, agent, { version, enabled }) {
       // Save the parsed req body to be picked up by getContextFromRequest().
       request.raw.req.payload = request.payload;
     }
-    return isHapiGte17 ? reply.continue : reply.continue();
+    return reply.continue;
   }
 
   function onPreResponse(request, reply) {
@@ -182,7 +146,7 @@ module.exports = function (hapi, agent, { version, enabled }) {
       agent._instrumentation.setDefaultTransactionName('CORS preflight');
     }
 
-    return isHapiGte17 ? reply.continue : reply.continue();
+    return reply.continue;
   }
 
   return hapi;
