--- test/instrumentation/modules/hapi/basic.test.js
@@ -1,9 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and other contributors where applicable.
- * Licensed under the BSD 2-Clause License; you may not use this file except in
- * compliance with the BSD 2-Clause License.
- */
-
-'use strict';
-
-require('./shared')('@hapi/hapi');
