+++ test/instrumentation/modules/hapi/set-framework.test.js
@@ -12,7 +12,7 @@ const agent = require('../../../..').start({
 });
 
 var isHapiIncompat = require('../../../_is_hapi_incompat');
-if (isHapiIncompat('@hapi/hapi')) {
+if (isHapiIncompat()) {
   // Skip out of this test.
   console.log(
     `# SKIP this version of '@hapi/hapi' is incompatible with node ${process.version}`,
